+++ hutool-captcha/src/test/java/cn/hutool/captcha/CaptchaTest.java
@@ -1,5 +1,6 @@
 package cn.hutool.captcha;
 
+import org.junit.Assert;
 import org.junit.Ignore;
 import org.junit.Test;
 
@@ -7,54 +8,63 @@ import cn.hutool.core.lang.Console;
 
 /**
  * 直线干扰验证码单元测试
+ * 
  * @author looly
  *
  */
 public class CaptchaTest {
-	
+
+	@Test
+	public void lineCaptchaTest1() {
+		// 定义图形验证码的长和宽
+		LineCaptcha lineCaptcha = CaptchaUtil.createLineCaptcha(200, 100);
+		Assert.assertNotNull(lineCaptcha.getCode());
+		Assert.assertTrue(lineCaptcha.verify(lineCaptcha.getCode()));
+	}
+
 	@Test
 	@Ignore
-	public void lineCaptchaTest() {
-		
-		//定义图形验证码的长和宽
+	public void lineCaptchaTest2() {
+
+		// 定义图形验证码的长和宽
 		LineCaptcha lineCaptcha = CaptchaUtil.createLineCaptcha(200, 100);
-//		LineCaptcha lineCaptcha = new LineCaptcha(200, 100, 4, 150);
-		//图形验证码写出，可以写出到文件，也可以写出到流
+		// LineCaptcha lineCaptcha = new LineCaptcha(200, 100, 4, 150);
+		// 图形验证码写出，可以写出到文件，也可以写出到流
 		lineCaptcha.write("d:/line.png");
 		Console.log(lineCaptcha.getCode());
-		//验证图形验证码的有效性，返回boolean值
+		// 验证图形验证码的有效性，返回boolean值
 		lineCaptcha.verify("1234");
-		
+
 		lineCaptcha.createCode();
 		lineCaptcha.write("d:/line.png");
 		Console.log(lineCaptcha.getCode());
-		//验证图形验证码的有效性，返回boolean值
+		// 验证图形验证码的有效性，返回boolean值
 		lineCaptcha.verify("1234");
 	}
-	
+
 	@Test
 	@Ignore
 	public void circleCaptchaTest() {
-		
-		//定义图形验证码的长和宽
+
+		// 定义图形验证码的长和宽
 		CircleCaptcha captcha = CaptchaUtil.createCircleCaptcha(200, 100, 4, 20);
-//		CircleCaptcha captcha = new CircleCaptcha(200, 100, 4, 20);
-		//图形验证码写出，可以写出到文件，也可以写出到流
+		// CircleCaptcha captcha = new CircleCaptcha(200, 100, 4, 20);
+		// 图形验证码写出，可以写出到文件，也可以写出到流
 		captcha.write("d:/circle.png");
-		//验证图形验证码的有效性，返回boolean值
+		// 验证图形验证码的有效性，返回boolean值
 		captcha.verify("1234");
 	}
-	
+
 	@Test
 	@Ignore
 	public void ShearCaptchaTest() {
-		
-		//定义图形验证码的长和宽
+
+		// 定义图形验证码的长和宽
 		ShearCaptcha captcha = CaptchaUtil.createShearCaptcha(200, 100, 4, 4);
-//		ShearCaptcha captcha = new ShearCaptcha(200, 100, 4, 4);
-		//图形验证码写出，可以写出到文件，也可以写出到流
+		// ShearCaptcha captcha = new ShearCaptcha(200, 100, 4, 4);
+		// 图形验证码写出，可以写出到文件，也可以写出到流
 		captcha.write("d:/shear.png");
-		//验证图形验证码的有效性，返回boolean值
+		// 验证图形验证码的有效性，返回boolean值
 		captcha.verify("1234");
 	}
 }