+++ hutool-core/src/main/java/cn/hutool/core/collection/CopiedIterator.java
@@ -1,7 +1,7 @@
 package cn.hutool.core.collection;
 
-import java.util.ArrayList;
 import java.util.Iterator;
+import java.util.LinkedList;
 import java.util.List;
 
 /**
@@ -23,7 +23,7 @@ import java.util.List;
  */
 public class CopiedIterator<E> implements Iterator<E> {
 
-	private List<E> eleList = new ArrayList<>();
+	private List<E> eleList = new LinkedList<>();
 	private Iterator<E> listIterator;
 	
 	public static <V> CopiedIterator<V> copyOf(Iterator<V> iterator){