+++ hutool-core/src/main/java/cn/hutool/core/util/XmlUtil.java
@@ -2,25 +2,34 @@ package cn.hutool.core.util;
 
 import java.beans.XMLDecoder;
 import java.beans.XMLEncoder;
+import java.io.BufferedInputStream;
 import java.io.BufferedWriter;
+import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStream;
 import java.io.Reader;
 import java.io.StringWriter;
 import java.io.Writer;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
 
 import javax.xml.namespace.QName;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.transform.OutputKeys;
+import javax.xml.transform.Result;
+import javax.xml.transform.Source;
 import javax.xml.transform.Transformer;
 import javax.xml.transform.TransformerFactory;
 import javax.xml.transform.dom.DOMSource;
 import javax.xml.transform.stream.StreamResult;
 import javax.xml.xpath.XPath;
+import javax.xml.xpath.XPathConstants;
 import javax.xml.xpath.XPathExpressionException;
 import javax.xml.xpath.XPathFactory;
 
@@ -70,7 +79,13 @@ public class XmlUtil {
 			// ignore
 		}
 
-		return readXML(new InputSource(file.toURI().toASCIIString()));
+		BufferedInputStream in = null;
+		try {
+			in = FileUtil.getInputStream(file);
+			return readXML(in);
+		} finally {
+			IoUtil.close(in);
+		}
 	}
 
 	/**
@@ -202,32 +217,19 @@ public class XmlUtil {
 	 * @return XML字符串
 	 */
 	public static String toStr(Document doc) {
-		return toStr(doc, null);
+		return toStr(doc, true);
 	}
 
 	/**
 	 * 将XML文档转换为String<br>
-	 * 此方法会修改Document中的字符集
+	 * 字符编码使用XML文档中的编码，获取不到则使用UTF-8
 	 * 
 	 * @param doc XML文档
-	 * @param charset 自定义XML文件的编码，如果为{@code null} 读取XML文档中的编码，否则默认UTF-8
-	 * @return XML字符串
-	 */
-	public static String toStr(Document doc, String charset) {
-		return toStr(doc, charset, true);
-	}
-
-	/**
-	 * 将XML文档转换为String<br>
-	 * 此方法会修改Document中的字符集
-	 * 
-	 * @param doc XML文档
-	 * @param charset 自定义XML文件的编码，如果为{@code null} 读取XML文档中的编码，否则默认UTF-8
 	 * @param isPretty 是否格式化输出
 	 * @return XML字符串
 	 * @since 3.0.9
 	 */
-	public static String toStr(Document doc, String charset, boolean isPretty) {
+	public static String toStr(Document doc, boolean isPretty) {
 		final StringWriter writer = StrUtil.getWriter();
 		try {
 			write(doc, writer, isPretty);
@@ -235,6 +237,25 @@ public class XmlUtil {
 			throw new UtilException(e, "Trans xml document to string error!");
 		}
 		return writer.toString();
+	}
+	
+	/**
+	 * 将XML文档转换为String<br>
+	 * 字符编码使用XML文档中的编码，获取不到则使用UTF-8
+	 * 
+	 * @param doc XML文档
+	 * @param isPretty 是否格式化输出
+	 * @return XML字符串
+	 * @since 3.0.9
+	 */
+	public static String toStr(Document doc, String charset, boolean isPretty) {
+		final ByteArrayOutputStream out = new ByteArrayOutputStream();
+		try {
+			write(doc, out, charset, isPretty);
+		} catch (Exception e) {
+			throw new UtilException(e, "Trans xml document to string error!");
+		}
+		return out.toString();
 	}
 
 	/**
@@ -275,17 +296,46 @@ public class XmlUtil {
 	/**
 	 * 将XML文档写出
 	 * 
-	 * @param doc {@link Document} XML文档
+	 * @param node {@link Node} XML文档节点或文档本身
 	 * @param writer 写出的Writer，Writer决定了输出XML的编码
 	 * @param isPretty 是否格式化输出
 	 * @since 3.0.9
 	 */
-	public static void write(Document doc, Writer writer, boolean isPretty) {
+	public static void write(Node node, Writer writer, boolean isPretty) {
+		transform(new DOMSource(node), new StreamResult(writer), null, isPretty);
+	}
+	
+	/**
+	 * 将XML文档写出
+	 * 
+	 * @param node {@link Node} XML文档节点或文档本身
+	 * @param out 写出的Writer，Writer决定了输出XML的编码
+	 * @param charset 编码
+	 * @param isPretty 是否格式化输出
+	 * @since 4.0.8
+	 */
+	public static void write(Node node, OutputStream out, String charset, boolean isPretty) {
+		transform(new DOMSource(node), new StreamResult(out), charset, isPretty);
+	}
+	
+	/**
+	 * 将XML文档写出
+	 * 
+	 * @param source 源
+	 * @param result 目标
+	 * @param charset 编码
+	 * @param isPretty 是否格式化输出
+	 * @since 4.0.9
+	 */
+	public static void transform(Source source, Result result, String charset, boolean isPretty) {
 		final TransformerFactory factory = TransformerFactory.newInstance();
 		try {
 			final Transformer xformer = factory.newTransformer();
 			xformer.setOutputProperty(OutputKeys.INDENT, isPretty ? "yes" : "no");
-			xformer.transform(new DOMSource(doc), new StreamResult(writer));
+			if(StrUtil.isNotBlank(charset)) {
+				xformer.setOutputProperty(OutputKeys.ENCODING, charset);
+			}
+			xformer.transform(source, result);
 		} catch (Exception e) {
 			throw new UtilException(e, "Trans xml document to string error!");
 		}
@@ -296,10 +346,10 @@ public class XmlUtil {
 	 * 创建XML文档<br>
 	 * 创建的XML默认是utf8编码，修改编码的过程是在toStr和toFile方法里，既XML在转为文本的时候才定义编码
 	 * 
-	 * @param rootElementName 根节点名称
 	 * @return XML文档
+	 * @since 4.0.8
 	 */
-	public static Document createXml(String rootElementName) {
+	public static Document createXml() {
 		final DocumentBuilderFactory dbf = DocumentBuilderFactory.newInstance();
 		DocumentBuilder builder = null;
 		try {
@@ -308,6 +358,19 @@ public class XmlUtil {
 			throw new UtilException(e, "Create xml document error!");
 		}
 		final Document doc = builder.newDocument();
+
+		return doc;
+	}
+	
+	/**
+	 * 创建XML文档<br>
+	 * 创建的XML默认是utf8编码，修改编码的过程是在toStr和toFile方法里，既XML在转为文本的时候才定义编码
+	 * 
+	 * @param rootElementName 根节点名称
+	 * @return XML文档
+	 */
+	public static Document createXml(String rootElementName) {
+		final Document doc = createXml();
 		doc.appendChild(doc.createElement(rootElementName));
 
 		return doc;
@@ -464,6 +527,45 @@ public class XmlUtil {
 	public static XPath createXPath() {
 		return XPathFactory.newInstance().newXPath();
 	}
+	
+	/**
+	 * 通过XPath方式读取XML节点等信息<br>
+	 * Xpath相关文章：https://www.ibm.com/developerworks/cn/xml/x-javaxpathapi.html
+	 * 
+	 * @param expression XPath表达式
+	 * @param source 资源，可以是Docunent、Node节点等
+	 * @return 匹配返回类型的值
+	 * @since 4.0.9
+	 */
+	public static Element getElementByXPath(String expression, Object source) {
+		return (Element) getNodeByXPath(expression, source);
+	}
+	
+	/**
+	 * 通过XPath方式读取XML的NodeList<br>
+	 * Xpath相关文章：https://www.ibm.com/developerworks/cn/xml/x-javaxpathapi.html
+	 * 
+	 * @param expression XPath表达式
+	 * @param source 资源，可以是Docunent、Node节点等
+	 * @return NodeList
+	 * @since 4.0.9
+	 */
+	public static NodeList getNodeListByXPath(String expression, Object source) {
+		return (NodeList) getByXPath(expression, source, XPathConstants.NODESET);
+	}
+	
+	/**
+	 * 通过XPath方式读取XML节点等信息<br>
+	 * Xpath相关文章：https://www.ibm.com/developerworks/cn/xml/x-javaxpathapi.html
+	 * 
+	 * @param expression XPath表达式
+	 * @param source 资源，可以是Docunent、Node节点等
+	 * @return 匹配返回类型的值
+	 * @since 4.0.9
+	 */
+	public static Node getNodeByXPath(String expression, Object source) {
+		return (Node) getByXPath(expression, source, XPathConstants.NODE);
+	}
 
 	/**
 	 * 通过XPath方式读取XML节点等信息<br>
@@ -528,7 +630,147 @@ public class XmlUtil {
 		}
 		return sb.toString();
 	}
+	
+	/**
+	 * XML格式字符串转换为Map
+	 *
+	 * @param xmlStr XML字符串
+	 * @return XML数据转换后的Map
+	 * @since 4.0.8
+	 */
+	public static Map<String, Object> xmlToMap(String xmlStr) {
+		return xmlToMap(xmlStr, new HashMap<String, Object>());
+	}
+	
+	/**
+	 * XML格式字符串转换为Map
+	 *
+	 * @param node XML节点
+	 * @return XML数据转换后的Map
+	 * @since 4.0.8
+	 */
+	public static Map<String, Object> xmlToMap(Node node) {
+		return xmlToMap(node, new HashMap<String, Object>());
+	}
+
+	/**
+	 * XML格式字符串转换为Map<br>
+	 * 只支持第一级别的XML，不支持多级XML
+	 *
+	 * @param xmlStr XML字符串
+	 * @param result 结果Map类型
+	 * @return XML数据转换后的Map
+	 * @since 4.0.8
+	 */
+	public static Map<String, Object> xmlToMap(String xmlStr, Map<String, Object> result) {
+		final Document doc = parseXml(xmlStr);
+		final Element root = getRootElement(doc);
+		root.normalize();
+		
+		return xmlToMap(root, result);
+	}
+	
+	/**
+	 * XML节点转换为Map
+	 *
+	 * @param node XML节点
+	 * @param result 结果Map类型
+	 * @return XML数据转换后的Map
+	 * @since 4.0.8
+	 */
+	public static Map<String, Object> xmlToMap(Node node, Map<String, Object> result) {
+		if(null == result) {
+			result = new HashMap<>();
+		}
+		
+		final NodeList nodeList = node.getChildNodes();
+		final int length = nodeList.getLength();
+		Node childNode;
+		Element childEle;
+		for (int i = 0; i < length; ++i) {
+			childNode = nodeList.item(i);
+			if (isElement(childNode)) {
+				childEle = (Element) childNode;
+				result.put(childEle.getNodeName(), childEle.getTextContent());
+			}
+		}
+		return result;
+	}
+
+	/**
+	 * 将Map转换为XML格式的字符串
+	 *
+	 * @param data Map类型数据
+	 * @return XML格式的字符串
+	 * @since 4.0.8
+	 */
+	public static String mapToXmlStr(Map<?, ?> data, String rootName) {
+		return toStr(mapToXml(data, rootName));
+	}
+	
+	/**
+	 * 将Map转换为XML
+	 *
+	 * @param data Map类型数据
+	 * @return XML
+	 * @since 4.0.9
+	 */
+	public static Document mapToXml(Map<?, ?> data, String rootName) {
+		final Document doc = createXml();
+		final Element root = appendChild(doc, rootName);
+		
+		mapToXml(doc, root, data);
+		return doc;
+	}
+	
+	/**
+	 * 给定节点是否为{@link Element} 类型节点
+	 * @param node 节点
+	 * @return 是否为{@link Element} 类型节点
+	 * @since 4.0.8
+	 */
+	public static boolean isElement(Node node) {
+		return (null == node) ? false : Node.ELEMENT_NODE == node.getNodeType();
+	}
+	
+	/**
+	 * 在已有节点上创建子节点
+	 * @param node 节点
+	 * @param tagName 标签名
+	 * @return 子节点
+	 * @since 4.0.9
+	 */
+	public static Element appendChild(Node node, String tagName) {
+		Document doc = (node instanceof Document) ? (Document)node : node.getOwnerDocument();
+		Element child = doc.createElement(tagName);
+		node.appendChild(child);
+		return child;
+	}
+
 	// ---------------------------------------------------------------------------------------- Private method start
+	/**
+	 * 将Map转换为XML格式的字符串
+	 *
+	 *@param doc {@link Document}
+	 *@param element 节点
+	 * @param data Map类型数据
+	 * @since 4.0.8
+	 */
+	private static void mapToXml(Document doc, Element element, Map<?, ?> data) {
+		Element filedEle;
+		Object value;
+		for (Entry<?, ?> entry : data.entrySet()) {
+			filedEle = doc.createElement(entry.getKey().toString());
+			element.appendChild(filedEle);
+			value = entry.getValue();
+			if(value instanceof Map) {
+				mapToXml(doc, filedEle, (Map<?, ?>)value);
+				element.appendChild(filedEle);
+			} else {
+				filedEle.appendChild(doc.createTextNode(value.toString()));
+			}
+		}
+	}
 	// ---------------------------------------------------------------------------------------- Private method end
 
 }