+++ hutool-cron/src/main/java/cn/hutool/cron/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * 定时任务模块，提供类Crontab表达式的定时任务，实现参考了Cron4j，同时可以支持秒级别的定时任务定义和年的定义（同时兼容Crontab、Cron4j、Quartz表达式）
+ * 
+ * @author looly
+ *
+ */
+package cn.hutool.cron;