+++ hutool-http/src/main/java/cn/hutool/http/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Hutool-http针对JDK的HttpUrlConnection做一层封装，简化了HTTPS请求、文件上传、Cookie记忆等操作，使Http请求变得无比简单。
+ * 
+ * @author looly
+ *
+ */
+package cn.hutool.http;