+++ hutool-log/src/main/java/cn/hutool/log/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Hutool-log只是一个日志的通用门面，功能类似于Slf4j。根据加入ClassPath中的jar包动态检测日志实现的方式，使日志使用个更加便利灵活。
+ * 
+ * @author looly
+ *
+ */
+package cn.hutool.log;