+++ hutool-setting/src/main/java/cn/hutool/setting/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Setting模块主要针对Properties文件读写做封装，同时定义一套自己的配置文件规范，实现兼容性良好的配置工具。
+ * 
+ * @author looly
+ *
+ */
+package cn.hutool.setting;